+++ internal/impl/azure/integration_test.go
@@ -102,6 +102,41 @@ input:
 		)
 	})
 
+	t.Run("blob_storage_streamed", func(t *testing.T) {
+		template := `
+output:
+  azure_blob_storage:
+    blob_type: BLOCK
+    container: $VAR1-$ID
+    max_in_flight: 1
+    path: $VAR2/${!count("$ID")}.txt
+    public_access_level: PRIVATE
+    storage_connection_string: $VAR3
+
+input:
+  azure_blob_storage:
+    container: $VAR1-$ID
+    prefix: $VAR2
+    storage_connection_string: $VAR3
+    targets_input:
+      azure_blob_storage:
+        container: $VAR1-$ID
+        prefix: $VAR2
+        storage_connection_string: $VAR3
+      processors:
+        - mapping: 'root.name = @blob_storage_key'
+`
+		integration.StreamTests(
+			integration.StreamTestOpenCloseIsolated(),
+			integration.StreamTestStreamIsolated(10),
+		).Run(
+			t, template,
+			integration.StreamTestOptVarOne(dummyContainer),
+			integration.StreamTestOptVarTwo(dummyPrefix),
+			integration.StreamTestOptVarThree(connString),
+		)
+	})
+
 	t.Run("blob_storage_append", func(t *testing.T) {
 		template := `
 output: